Bug#31018950 NDB_CONDITION_PUSHDOWN.TEST FAILS AFTER PATCH FOR BUG#30350206 - CONCAT() ...
Generating the condition pushdown code relied on buggy behaviour
of Item::save_in_field() which allowed strings longer than the
max length of the Field to be saved.

This has been corrected by the patch for bug#BUG#30350206, which
in turn caused a test case in ndb_condition_pushdown.test
to start failing.

Since Mysql-8.0 the particular code in the condition pushdown handler
has been replaced with (simpler) code not depending on ::save_in_field().

This patch pack port the relevant part (patch 1/5) of the patch
which did that refactoring. The original commit comment for
that patch is pasted below:

------------------------------------

Bug#29703986 CLEANUP OF HA_NDBCLUSTER_COND.CC CODE

Simplify code for creating pushed condition NdbScanFilters for
LIKE / NOT LIKE condition.

The existing code is rather obfuscated by:
- Depending on 'length' of the pattern to match, we either extract the pattern
  value from the value provided by save_in_field(), or directly from the Item
  expression defining the value.
- The pattern value extraction is further depending on whether the value 'is_const'
  or 'is_cached'.
- Same goes for the 'length' of the pattern.
- The 'pattern' and 'length' extraction for NOT LIKE is slightly different
  from LIKE

... everything without a single comment describing the logic behind this.

Diving further into the implementation of the Ndb_item get'ters, used to lookup
the pattern values, we find that there even are cases where we return a NULL
string for the pattern value. (If Ndb_item is a NDB_VALUE, and 'type' not a STRING_ITEM).

.. still without comments.

The entire semantics around how the pattern value used for the LIKE filters
are set is rather unclear, and feels unsafe.

Looking at how Item_func_like::val_int() (The server side evaluation of LIKE)
extract the same pattern value, we find that simply calling Item::val_str() is
always sufficient.

This patch reimplement the 'pattern match' value extraction for NDB_LIKE_FUNC
and NDB_NOTLIKE_FUNC with similar logic as used in Item_func_like::val_int().

Reviewed by: Arnab Ray <arnab.r.ray@oracle.com>