Bug#30434663 NDB_IMPORT FAILS WHERE MYSQLIMPORT SUCCEEDS
mysqlimport accepts a trailing field terminator at the end of a line;
ndb_import does not.

This is not a parser issue, because the parser does not know how
many fields a line should have, and it does not distinguish a
trailing field separator from an empty final field. So, handle
trailing terminators when evaluating parsed lines.

This version of the patch adds logic to return the final field
to the free list at evaluation time.

Approved-By: Maitrayi Sabaratnam <maitrayi.sabaratnam@oracle.com>