Bug#28716103 - MULTI-TABLE UPDATE OF PRIMARY KEY RESULTS IN ERROR 1032 "CAN'T FIND RECORD"
Issue:
In a multi-table-update statement, if the key of the first table is
being updated, a temporary table execution strategy is chosen instead of
on-the-fly-update (even if it is the only table being updated).

The join with the other tables of the update statement can produce
duplicate entries into the temporary table. Duplicates are prevented by
marking the field having the rowids (primary-key) of the main table as
unique, if the length of the field is not long (<512 bytes). Otherwise a
new field (TABLE::hash_field) which will hold the hash of the rowid is
added to the temp table and a unique constraint is created on this
field.

The unique constraint does not get enforced as the hash field value is
not updated. These duplicate entries of the temporary table cause the
error 1032(can't find the record) when the rows of the main table are
tried for update.

Fix:
For every new record to be inserted into the temp table update the
hash_field, check if there is a collision with an existing entry and
ignore if duplicate.

Allocate the ORDER("group") structure on the heap. This is referenced in
the "tmp_table->group" field and is used to find the rowid field on
which the hash is to be calculated.

Change-Id: I03f4d38660ec35848c23e6376a9820bac21bfcf6