Bug #31037263 MEB ON WINDOWS IS DOA IF BUILT WITH CURL USING ZLIB 1.2.11
The name of the zlib DLL had changed from zlib1.dll in 1.2.5 to
zlib.dll (without '1') in 1.2.11. This lead to no zlib library being
bundled in the install zip. WHen mysqlbackup.exe was started, it would
immediately fail.

Fixed by looking for either zlib.dll or zlib1.dll in dependencies of
libcurl.dll. We need to accept either in order not to introduce sudden
build breaks while upgrading. After this change has been pushed, we
can rebuild the 3rd party curl with the new zlib.

Change-Id: I38605274af2183af2caa8e94d09e6bb3e771978b