WL#13425: implementation.
Since WL#13517 has been pushed to trunk, we already offload to RAPID
the queries which have semijoins operations (as converted by MySQL
from IN(subquery) or EXISTS(subquery) predicates). MySQL creates a
semijoin operation for such predicates as long as the subquery has no
aggregation, and is linked to the root of the WHERE clause through a
chain of AND (there are other conditions, not necessary to discuss
here). Thus, we do not offload when there is:

  - IN/EXISTS(subquery) OR other_condition
  - IN/EXISTS(subquery with aggregation e.g. COUNT(*))
  - NOT IN/EXISTS(subquery) (a.k.a. antijoin ; while MySQL can
    transform NOT IN and NOT EXISTS to anti-join, it's not
    offload-able in this form, yet).

The goal of this WL is to transform the not offloaded patterns above
into joins to derived tables, which is something which RAPID
supports.

Example statement (case 1):

 SELECT *
 FROM ot
 WHERE oe1 IN (SELECT ie1 FROM it1) OR
       oe2 IN (SELECT ie2 FROM it2)

will internally be rewritten to:

 SELECT ot.*
 FROM ot LEFT JOIN (SELECT DISTINCT ie1 FROM it1) AS der1
         ON oe1=der1.ie1
         LEFT JOIN (SELECT DISTINCT ie2 FROM it2) AS der2
         ON oe2=der2.ie2
 WHERE der1.ie1 IS NOT NULL OR der2.ie2 IS NOT NULL

Another example to handle is (case 2):

 SELECT *
 FROM ot
 WHERE oe1 IN (SELECT COUNT(ie1) FROM it1);

which will be rewritten to

 SELECT ot.*
 FROM ot JOIN (SELECT COUNT(ie1) AS Name_exp_1 FROM it1) AS derived_1_2
         ON oe1=derived_1_2.Name_exp_1

(there is no need to add DISTINCT to the subquery in this case).

We will also handle this antijoin (case 3):

 SELECT *
 FROM ot
 WHERE oe1 NOT IN (SELECT COUNT(ie1) FROM it1);

we will rewrite to:

 SELECT ot.*
 FROM ot LEFT JOIN (SELECT DISTINCT COUNT(ie1) AS Name_exp_1 FROM it1) AS derived_1_2
         ON oe1=derived_1_2.Name_exp_1
 WHERE derived_1_2.Name_exp_1 IS NULL

Decorrelation of equalities will be done, if the subquery is not aggregated:

 SELECT *
 FROM ot
 WHERE EXISTS (SELECT * FROM it1 WHERE oe1=ie1) OR other_cond;

will be rewritten to:

 SELECT ot.*
 FROM ot LEFT JOIN (SELECT DISTINCT ie1 AS Name_exp_1 FROM it1) AS derived_1_2
         ON oe1=derived_1_2.Name_exp_1
 WHERE derived_1.2.Name_exp_1 IS NOT NULL OR other_cond;

Same for IN. For NOT EXISTS and NOT IN, we'll do the same, and we'll
also decorrelate "a<>b" (and >,>=,<,<=):

 SELECT *
 FROM ot
 WHERE NOT EXISTS (SELECT * FROM it1 WHERE oe1<>ie1)

will be rewritten to:

 SELECT ot.*
 FROM ot LEFT JOIN (SELECT DISTINCT ie1 AS Name_exp_1 FROM it1) AS derived_1_2
         ON oe1<>derived_1_2.Name_exp_1
 WHERE derived_1_2.Name_exp_1 IS NULL;

The entry point to inspect this code is
SELECT_LEX::transform_table_subquery_to_join_with_derived().

Reworked a bit the tracing of transform_scalar_subquery_to_derived().

In Item_in_subselect::fix_fields():
if (exec_method == SubqueryExecMethod::EXEC_SEMI_JOIN)
    return !((*ref) = new Item_func_true());
was dead code (if we have chosen semi join, the subquery has
permanently disappeared, we cannot come to fix_fields anymore).

Item_subselect::update_used_tables() is simplified and so
upper_select_const_tables() is now useless and removed.

SELECT_LEX::can_skip_distinct() : it is a little helper function to
group some if() tests.

Moved alloc_qep() to before finalize_table_conditions() as the
latter can evaluate a subquery which may be handled with IN-to-EXISTS
and thus may need access to this subquery's qep_tab (see
calculate_subquery_executions()).

When resolving table value constructors: synchronized all_fields and
fields_list.

In opt trace tests: there were 2 blocks for semijoin decision: the
initial, not-sure decision (resolve_subquery()) which isn't printed
anymore; only the final one (in flatten_subquery()) is, now.

rapid.filesort_merge: test had randomness; stabilized it.

rapid.subquery: with the WL it fails; incorporated Cagri's patch to
fix it (a draft patch for BUG#31068290).

As transform_scalar_subquery_to_derived() and the new
transform_table_subquery_to_derived() share code, the numbering of
derived tables changed ("derived_%d_%d"), this is visible in tests
which have scalar subqueries changed to derived tables.

In secondary_engine.cost_threshold :
this query
EXPLAIN SELECT * FROM t1 tt
WHERE x IN (SELECT SUM(x) FROM t1 WHERE x = (SELECT tt.y));
doesn't use the MOCK secondary engine anymore. Because: it's
aggregated (=>no semijoin), we transform to derived, but can't
decorrelate (as it's aggregated), so bail out with error, which sends
the query to InnoDB. Whereas, pre-WL, with the MOCK engine, we
directly do IN-2-EXISTS, and the MOCK engine accepts that.

With this WL we get new offloads in RAPID tests: filesort_merge,
group_min_max, having, i_filesort, i_subquery,
i_subquery_mat_cost_based, olap, opt_hints_subquery, semijoin (and for
not-offloaded ones, the reason changes), subquery, subquery_mat,
subquery_sj, sum_distinct, type_temporal_fractional.

Reviewed By : Dag Wanvik <dag.wanvik@oracle.com>