Bug #31069510: REMOVE UNNEEDED SUBQUERY EXISTS OPTIMIZATION
There is an optimization that tries to rewrite EXISTS ( SELECT a, b ... )
into EXISTS ( SELECT 1, 1 ... ), since EXISTS does not care about the
contents of the fields.

However, it only modifies fields_list, not all_fields, and most temporary table
decisions are based on all_fields. (This breaks the relationship between
fields_list and all_fields, which is that the former is a sublist of the
latter.) Furthermore, the entire optimization is moot, since
Query_result_exists_subquery::send_data() doesn't ever actually evaluate its
items. Thus, remove it.

Change-Id: Id78aeec8c38a15e8f224d7232a0ecd968b77da09