WL#13796: Rewrite 'skip-handling' in ::prepareResultSet()
NdbResultSet::prepareResultSet() is used when complete batches of
result rows has been received from the SPJ node(s) expected to
return results. It will establist parent child relationship
between the result rows. As part of this it will also eliminate
rows which did not meet the full-join, outer-join, semi-join or firstMatch
propertes as specified.

Previous implementation was based on filling in a bitmask of
child nodes having rows matching 'this' row. When all children
for a spefic row had been checked, we compared the match-bitmask
against the required matches needed for a 'pass'.

This doesn't really fit the antijoin semantics planned for WL#13796 -
For an antijoin a 'match' is actually a 'skip'

This patch change this algorithm such that a match-mask is
no longer collected for each tuple. Instead we employ the
full-join, outer-join, semi-join or firstMatch semantics as part
locating matches from each children, and try to 'skip' as
soon as we can decide that the row fails to fullfill the
specified join semantics.

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>