WL#13796: Let SPJ API eliminate matching rows from antijoins
Introduce usage of the existing (not previously used) MatchType 'MatchNullOnly'.
SPJ handler integration will set this flag for anti-join operations where the
entire anti-joined queryTree-branch is completely evaluate by the SPJ block.

That is:
  - The entire anti-joined query-branch has been pushed down
  - There are no unpushed conditions in this branch.

Furthermore, antijoin-elimination of matching rows is implemented in
::prepareResultSet().

As only the 'firstMatch' is required from the anti-joined queryTree-branch,
this patch will also set firstMatch result handling in the NdbResultStreams
associated to any queryTreeNodes *below* the head of the anti-joined branch.

The net effect of this patch is that fewer rows are presented to
the MySQL server from the NDB handler interface, thus reducing required
MySQL server overhead needed to produce the final result set.

If we can't set the 'MatchNullOnly' MatchType, the antijoin result set
will still be returned to the MySQL server as an full outer join.
The SPJ block itself will also still return a full outer joined
result set for the anti join.

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>