WL#13796: Let SPJ block do a LQH-firstMatch REQ for antijoins
Patch introduce and set a NI_ANTI_JOIN-flag as part of the queryTree
requestInfo bits sent to the SPJ block. As an antijoin is a stricter
variant of firstMatch, the SPJ block will setFirstMatchFlag() on
ScanFragReq's having the NI_ANTI_JOIN-flag set (Leaf nodes only)

Furthermore: If a queryTreeNode is part of an antijoin, that is:
'Having a antijoin as an ancestor treeNode', the entire queryTree-branch
below the antijoin only need to evaluate the 'firstMatch'.

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>