WL#13978: Allow 'non-dependent' scalar subqueries to be part of pushed conditions
Patch remove the exclusion of conditions having a used_tables()
set containing the INNER_TABLE_BIT from being part of a pushed
conditions.

That let a 'non-dependent' scalar-subquery to pass through to
the ndb_serialize code in the condition pushdown handler.

That code was already prepared to identify conditon parts with had
no dependencies on other tables, thus could be treated as constant
expressions and preevaluated into a const-value in the ScanFilter.

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>