WL#13978: Fix explain tree-format to include subqueries found in pushed-conditions
The tree-explain code will 'walk' any filter conditions attached
to the FilterIterators and look for any subqueries being part of
the 'filter'. Any such found subqueries will be reported as
a childer() of the FilterIterator and thus explained as a
descendant-Iterator.

However, any condition terms being part of a pushed condition, is
removed from the original (filter-)conmdition. Thus the subqueries
now being converted into a pushed condition by WL#13978, was not
included in the explained tree-format execution plan.

Thus patch fixes that by implementing a 'children()' method for the
TableRowIterator base-class, similar to the existing
FilterIterator::children() method. All NDB-table Iterators
possibly taking a pushed_condition will be derived from the
TableRowIterator. Thus searching for subqueries in the
pushed_condition here will be sufficient.

Reviewed by: Mikael Ronstrom <mikael.ronstrom@oracle.com>