WL#13341: Store options for secondary engines.
Two new attributes, ENGINE_ATTRIBUTE and SECONDARY_ENGINE_ATTRIBUTE
are introduced for tables, columns, indices and tablespaces.
The value must be a string literal which is either an empty string or
valid json.

The attribute can be altered using the appropriate ALTER
command. Attribute values cannot be updated, only replaced.

Attributes are stored in the data dictionary and can be retreived from
the corresponding dd cache object or the TABLE_SHARE. Attributes are
displayed by SHOW CREATE TABLE (unless they are empty).

The motivation is to provde a generic way to supply storage engine
specific attributes, and to let storage engine developers add new
attributes without having to make changes to the parser/server code.

Change-Id: I730100fd8e4a148d6175154584ce94508a17e6ad