Bug#30613342 MTR TEST NDB_ALTER_TABLE_BACKUP CAN FAIL
Problem:
--------

The ndb_alter_table_backup test failed sporadically due to errors
in the output log of the second MySQL server. The test involves
restoring a backup containing a table with old temporal data types.
Discovery of this table is expected to fail since support for the
above data types has been removed in 8.0 and there already exist
suppressions of errors in the output log.

The additional error messages are a result of a timing issue where
schema synchronization of the second MySQL server occurs after the
backup is restored. Schema synchronization attempts to create this
table in the DD which results in the additional error messages
being spewed out

Fix:
----

Suppress the expected error messages

Change-Id: I82704e0ef25f2c9488d1737da35846f4b18c0956