Bug#31100818 MTR: REPLACE_RESULT IS NOT ABLE TO REPLACE LONG STRINGS IN PATH NAME
Description:
------------
This is a backport of the fix made for Bug#26630826 in 8.0.

The mysqltest command 'exec' executes the shell command using popen
function which returns the pointer to a file containing the output of
the command. Then 512 bytes per chunk is read from that file, and the
pattern(s) mentioned in the previous replace_result command is
searched. However, this might fail if the pattern gets truncated when
a chunk of 512 bytes is read.

Fix:
----
Instead of reading 512 bytes per chunk, read the entire line from the
file and use it to match the patterns.

Change-Id: Iba8151834814f1d7b754b840c7fa24fc16c9eab1