Bug #31001732 REFACTOR LOCK-SYS TO USE PAGE_ID_T WHERE APPLICABLE
Temporary post-push fix.
For some reason both Valgrind and ASAN release builds fail to initialize the m_fold member of the RecID struct correctly, which leads to enqueueing the lock in wrong bucket of hash-table and to crash when we try to dequeue the lock later on from the correct bucket (and find it empty).

I'm still investigating what is the real root cause, but what works as a temporary bandaid is to change the type of lock_rec_fold function to accept the page_id_t argument by value instead of by reference.
The problem is somehow related to the issue that page_id_t is small enough (8 bytes) to fit into a single register, so to pass it by reference the compiler must store the register content on stack and pass the correct stack address to the function - somehow both Valgrind and ASAN builds fail at computing the address correctly - I'm still not sure why.
However, if the value is passed by value, then it seems to work correctly as the register is simply passed as is.

Note that changing const page_id_t& to const page_id_t is something we might want to do in many places anyway, given that the size of the struct is comparable to size of the pointer to it, and perpahs involves less bookkeeping (and less bugs?) so I believe this is not a bad change.
And I don't see how this change could make things worse.

Reviewed-by: Marcin Babij <marcin.babij@oracle.com>
RB:24189