BUG#31290638 - USE VARLEN ENCODING BY DEFAULT IN THE MTR RAPID SUITE
Add combinations file to make entire rapid suite run with both
varlen and dictionary encoding.

If tests have to be run with the default encoding, run MTR
with --skip-combinations option ( *_dict and *_varlen tests always run).
This will be added for the per-push runs too.

If tests have to be run with both varlen and dictionary
encodings, remove the --skip-combinations option.
The daily runs will run with the combinations to ensure
more coverage is attained.

Change-Id: Id08e1ada70d16b340cc2c0c77ee8f84718858afb