Bug #30774226 SPECIAL SCANTABREQ DISTRIBUTION KEY LEADING TO DBTC CRASH
A SCANTABREQ with a distribution key value of 0xffffffff causes DBTC
to crash, due to incorrect handling of the distribution key value.

A testcase is added, using error insert to force this value.
  testPartitioning -n startTransactionHint_orderedIndex_MaxKey

The logic error in DBTC is fixed to avoid the problem.

Approved by : Ole John Aske <ole.john.aske@oracle.com>