Bug#28240054 - SIG11 IN SELECT_LEX_UNIT::OUTER_SELECT() AT SQL/SQL_LEX.H
Problem:
Removal of a subquery having a "group by" operation, which is
used by an outer select results in server exit.

In a query, the "group by" subquery can be considered redundant and
removed assuming as there are no aggregate functions. But if the
subquery is used by the outer context, it should not be removed.

Fix:
Do not remove the subquery if it is needed for the select list.

Change-Id: I696461cb59734984fd7f7095be161999461470e1