Bug#30674616 - UPDATE QUERY FAILS WITH "ERROR 1071 (42000): SPECIFIED KEY WAS TOO LONG; MAX KEY LENGTH IS 1024 BYTES"
Issue:

In case of an update statement involving multiple tables, an internal
temp table is used to hold the rows before updating the target table.

In the temp table, a fixed length string field(Field_string) is used to
store the target tables' primary key values of the rows qualifying for
the update. Later these values from the temp table are used to identify
and update the rows of the target table.

When a temp table grows big and no longer fits in the temp storage
engine table, it is moved to the on-disk storage(InnoDB). InnoDB expects
all fixed length string fields to be less than 1024 bytes (struct
dict_field_t::fixed_len uses 10 bits to store the length).

If the fixed length string field, in the temp table, is bigger than 1024
bytes, an error is returned while migrating the temp table to InnoDB.

Fix:

Use variable length string field (Field_varstring), in the temp table,
to hold the primary key values of the target table.

Change-Id: Ife7fecf1ff41240016dfd57b958341e870e9996a