Bug#30876203 INTERNAL OCIADMIN USER SESSIONS CAN BE KILLED BY NORMAL USER
Description:
-----------
If rapid server is killed for any reason then rapid plugin creates user session
for auto recovery. It creates the session as following :
1. Create a new session using session services.
2. Get the thd from the session service.
3. Get the security_context handle from the thd using security_context services.
4. Update the security_context with the authid 'ociadmin'.
   (who has system_user privilege)

The problem here is that this connection has system_user privilege but another
connection that does not have system_user privilege who can kill the former.

It happens because WL#12364 expects a system_user flag in the THD.
At the moment that gets updated when security_context is set to the thd
using thd_set_security_context() API. But calling that API is optional
when user updates the security_context of the THD itself. Calling
thd_set_security_context() is required if user wants to set new
security_context to the thd .

Fix:
----
(1) This situation may happen to anyone who wants to fill the existing
security_context of thd using security_context_lookup() API. We know the
security_context already cross references the thd. Therefore, during the
lookup we can check if sctx references a thd, if so, update the flag in
that thd.

(2) We call set_system_user_flag() to update the flag in the thd. This api
intern calls Security_context::has_global_grant() that looks for current_thd
to acquire the ACL_cache_lock. It may happen that current_thd is not set
when we are constructing the thd e.g. calling thd_set_security_context().
Added following changes to fix this issue :
a. Updated the cross references of thd in the sctx.
b. Check for referenced thd and then current_thd in the
   Security_context::has_global_grant() API.

(3) In X Plugin, after the lookup, we were setting the flag in thd using
thd_set_security_context() API. That call is not required now.
Therefore removed it.

Testing
--------
1. x.system_user_kill mtr test covers this case.
2. Fix verified by rapid developer. Bug#31177903 is created to add test in
   rapid suite.
3. Itch run in mysql-8.0-itch in progress.
   push id: 16163136
   2020-04-14 13:08:40

Review:
-------
RB#24252