WL#13979: MySQL GCS: Reduce minimum value of group_replication_message_cache_size
This patch lowers the lower bound for the system variable
group_replication_message_cache_size.

Some group replication deployments are done on hosts that
have small memory and stable network. In addition, the
write workload running through the system is light.

On those scenarios, it may make sense to be able to set a
lower bound for the max message cache size.

Reviewed by: Nuno Carvalho <nuno.carvalho@oracle.com>
Reviewed by: Tiago Jorge <tiago.jorge@oracle.com>