Bug#29998457 - MYSQLDUMP GENERATES INVALID INSERT                STATEMENTS WITH VARBINARY COLUMNS
DESCRIPTION
===========
Binary and varbinary columns are now tagged with "_binary "
in the mysqldump output. However, the memory allocated for
the insert statement does not take the length of the
'_binary ' string into account. The resulting pointer
arithmetic causes invalid insert statements being generated
in the mysqldump output.

FIX
===
Account the memory for the "_binary " string that gets
added in front of the string while reallocating the memory.

RB: 24098