BUG#30395151: CREATE TEMPORARY TABLE OPERATION WITH FUNCTIONS RECORDED IN BIN LOG
DESCRIPTION:
============
With binlog_format set to MIXED, CREATE TEMPORARY TABLE
with functions are recorded in the binlog.

ANALYSIS:
=========
Fix for BUG#28258992 allowed binlogging of a function call
in STATEMENT format when in MIXED mode, if the function
contained operations on temporary tables. The fix allowed
binlogging of the function call in STATEMENT format based
on the conditions that the binlog format is MIXED and a
temporary table is accessed in the function context. The
consequence of these conditions was this issue.

When in MIXED mode, binlogging of functions containing
operations on temporary tables in STATEMENT format is
likely to break the replication. A simple example would be
creating a temporary table and then create a function to
perform a DML on that temporary table.  DDL on temporary
tables are not binlogged by default. Thus, binlogging the
function call in STATEMENT format in this case would cause a
failure on slave. The slave tries to execute all the
statements in the function. Since it doesn't have the
temporary table created, execution of DML will report an
error. Thus leading to break in Replication.

This issue also exists with triggers that have create or
drop temporary table operations in their sub-statements.
In case of stored procedures and events, each
sub-statement is binlogged individually and has been taken
care of, by the existing code. However, the issue exists
with stored functions invoking stored procedures that have
temporary table operations in their sub-statements.

FIX:
====
The fix marks statements with stored routines and triggers
containing create or drop temporary table operations in its
sub-statements as unsafe to be binlogged in STATEMENT
format. This ensures that such statements are binlogged in
ROW format when binlog format is MIXED. As part of the
patch, Fix for BUG#28258992 is reverted.

NOTE:
=====
This patch also fixes
BUG#30320009: ASSERTION TRX_IS_REGISTERED_FOR_2PC(TRX).
The fix for BUG#28258992 caused an assertion to be
triggered during 2PC. The assert condition is not hit after
reverting the patch for BUG#28258992.

RB: 23957
Change-Id: I555ba4d4ed60d1e16335d245954f419ff7317a04