BUG#31166792: GR: ASSERTION `MP->M_U.M_SAFE_PTR != NULLPTR' FAILED.
gr_parallel_start_uninstall forces a race condition between
  `UNINSTALL PLUGIN group_replication;`
and
  `START GROUP_REPLICATION;`
Being the `START` executed when the `UNINSTALL` is already ongoing.
Depending on thread scheduling, there is a small window on which
the START can still reach the plugin just before it is uninstalled,
leading to use of resources already deleted.

To avoid the above situation, now `START GROUP_REPLICATION` checks
if the plugin is being uninstalled, on which case it ends with the
appropriate error.

RB: 24251