BUG#31180464: THREADS MAY READ STALE VALUES OF GTID_MODE
Fix: Simplify gtid_mode implementation:

- Encapsulate gtid_mode in a class, implement it in it its own .cc
  file, instead of in the header and in rpl_gtid_misc.cc.
- Don't cache or version the value.
- Use atomics, not locks, when reading the value.

Extra: Minor improvements and refactorings:

- Created Checkable_rwlock::Guard, a RAII class to hold a read or
  write lock for the duration of a block.  Used it in only one file
  for now: rpl_group_replication.cc.  Later we can use it to simplify
  more code.
- Simplify prototypes for Gtid_state::begin_gtid_wait and
  get_server_startup_requirements: they don't need an argument to
  specify the lock type.
- Make gtid_mode_lock be statically allocated rather than
  heap-allocated, to simplify memory management.
- Don't use typelib in the enum-to-string helper functions.
- Remove obsolete code in get_master_version_and_clock. We no longer
  allow masters with gtid_mode set to invalid values that begin with
  ON or OFF.  This code was created in 5.6 which only had
  gtid_mode=ON|OFF, but we anticipated ON_PERMISSIVE|OFF_PERMISSIVE.
- Remove obsolete comment in mysqld.cc.
- Fix error message generated when failing to write
  Incident_log_event.  If the query is long, it now cuts the query
  instead of the message.
- Prefix members of Checkable_rwlock with m_.
- In the error message generated for SET gtid_mode=ON_PERMISSIVE when
  replicate-same-server-id is enabled, explain the problem more
  precisely and don't repeat the gtid_mode twice.

Background:

In 5.7, gtid_mode is protected by 4 locks. A thread that sets it needs
to hold all locks, and a thread that reads it needs to hold at least
one of the locks. gtid_mode is only updated seldom, so it was not a
problem to require many locks. Many places that read gtid_mode will
already hold one of the locks for other reasons, so for those places
the locking scheme was not a problem either.

However, there were a couple of places where no lock was needed except
in order to read gtid_mode. In 8.0 (WL#8599), we reduced locking in
those places. To accomplish that, we introduced a global atomic
integer which represents the version of gtid_mode. In addition, we
introduced the class Gtid_mode_copy, which caches the global gtid_mode
and its version number.  When a caller requests gtid_mode from an
instance of the class, the class first compares the cached version
number with the global version number; if the cache is out of sync, it
is refreshed; and finally it returns the cached value to the caller.

Problem:

Some cache objects are shared between multiple threads, but the
members of the cache class are not protected against concurrent
use. This can, at least theoretically, result in problems if threads
T1 and T2 compete to read the same cache at about the same time,
recently after some other thread has changed the global gtid_mode. If
T1 comes first, it will update the cache with the most recent version
number and gtid_mode. However, there is no guarantee that both those
values are visible to T2. In the worst case, T2 may see the new
version number and the old gtid_mode, thus it will get the wrong
result.

In addition, this scheme is over-complex. There is no need for
versioning, caches, or locks; we should just make the global gtid_mode
be an atomic.

RB: 24259
Approved-by: Pedro Figueiredo <pedro.figueiredo@oracle.com>
Approved-by: Pedro Gomes <pedro.gomes@oracle.com>