Bug#31128739: THREAD 140231812572928 ALREADY OWNS A LATCH DBLWR AT LEVEL 11
There is a call to buf_page_io_complete() in Fil_shard::do_io() which
for writes, will try get buf_pool->LRU_list_mutex. At this point, the
Double_write::m_mutex (SYNC_DBLWR) should not be held since it is a
lower level than SYNC_BUF_FLUSH_LIST. But it can be if the call to
do_io() comes through Double_write::enqueue().

This patch rearranges the priority of SYNC_DBLWR to raise it higher
than SYNC_BUF_FLUSH_LIST and similar latches.

It also cleans up dblwr::write() to use a local copy of the space_id
and makes sure that buf_page_io_complete() is called after the call
to Double_write::write_to_datafile() when sync is true even if the
return code is not DB_SUCCESS.

Approved by Yasufumi and Kuba in RB#24273