Bug #31158239 	CURRENT_TEST IS NOT LOGGED WHEN SERVER IS NOT RESTARTED BETWEEN TESTS ON WINDOWS
The fopen has a bug (submitted in: https://developercommunity.visualstudio.com/content/problem/921279/createprocess-does-not-inherit-fappend-flags-set-b.html )
that opens a file for append in a way that does not allow the append flag being duplicated, or inherited in a duplicated handle, which in turn is supplied to a child process that is being created.
This works around the issue reopening the output files for append always and correctly.

RB#24240