Bug #30852113 	SOME MTR TESTS FAIL WHEN RUN WITH --REPEAT=N (N>1), BECAUSE OF USE OF LOG SEARCH
Fixed the `assert_grep.inc` and `search_pattern.inc` scripts to reset the search results after "CURRENT_TEST: " line. This line is not to be found by accident in any tests, but is used by MTR to separate mysqld server log between test cases run on single worker. Also, there is a number of test cases that use these search scripts specifically on the log file.
The tests that used `assert_only_after` with "CURRENT_TEST: " have this line removed - it's not needed and sometimes it only had bugs like an additional space, testname misspelled, colon instead of dot.

RB#23783