Bug#31000043 - MINOR PERFORMANCE DEGRADATION AT SYSBENCH-POINT_SELECT TEST
Map of THD instance to Session instance is maintained by the
Srv_session module. Map is protected with a read write lock.
Every Session service does a sanity check that verifies Session
instance is present in the Map. This works well with the less number
of threads. With more thread accessing the read write lock minor
performance degradation is observed.

Session instance contains state of the session (i.e. CREATED,
ATTACHED, DETACHED and CLOSED). Session instance is added to the
Map when session is opened. And it is removed when session is
closed. Session instance in Map can be in ATTACHED or DETACHED
state. The state of the valid session is ATTACHED and DETACHED.
Session is valid even when opened and added to the list but not
attached yet. We do not have any state to represent Session in
opened but not in attached situation for now.

To fix the issue, added one more state "OPENED" for the Session.
This state is set when session is opened and after adding
Session instance to Map. Now instead of checking session existence
in Map, state of the Session instance is checked. If state of the
Session instance is OPENED, ATTACHED or DETACHED the it is a valid
session otherwise invalid.
For debug build, Map is still checked to make sure valid session
exists and invalid session doesn't exists in the opened sessions list.

Change-Id: If02727c818f30fd984347888fadb9954d3794716