Bug #31221808 CLEANUP ZLIB.CMAKE
The cmake code for checking zlib featurs was obsolete, remove it.

Also remove the "fallback to bundled" if "system" does not have the
right features: if the user asks for "system" and this lacks necessary
features, the build should fail.

Experimenting with misc zlib versions, it turns out we need 1.2.11
Add explicit version check.

Default for -DWITH_ZLIB is "bundled" which has version 1.2.11

Change-Id: Iad4fc19c7397477eb5eeef6a1ef7a5cd6661fd9c