WL#13988 ADD COVERAGE PARAMETER TO ATRT WHICH ENABLES GRACEFUL CLUSTER RESTART
POST PUSH FIX

Modifies autotest-run.sh to make "coverage" parameter accessible by
atrt.

Currently this parameter is not accessed by atrt as its been handled
in autotest-run.sh.

Coverage files should be gathered only for builds with coverage and
not for normal builds. This parameter is an indication that coverage
is enabled. Making this paramter accessible to atrt would make sure
that coverage files are gathered only for coverage builds.

Change-Id: Id7eadf3de4e48f6d22bf92030a7e4961eb0296c7