Bug#30139031: GET_PACKET_LENGTH() SHOULD RETURN ULONG
Description: There is an integer loss of precision in
a function during connection leading to a diverted
control flow.

Analysis: A function which is supposed to return ulong
returned unit which leads to loss of integer precision.
This lead to a diverted control flow.

Fix: Change the return value of the function to ulong
instead of uint.

RB#24097