Bug#31174834 Failover node not supporting correctly GR in routing_connection component test
Problem
=======

RouterRoutingConnectionMDUnavailableTest.IsConnectionKeptWhenClusterOverloaded
fails spuriously after the cluster primary node is detached from the cluster.
That occurs when Router is in the middle of the metadata refresh, in
which case the secondary nodes are responsible for the replicaset status
update. However this component test does not instrument the secondary
nodes with GR capabilities.

Fix
===

Instrument mock to support GR on the secondary nodes for the
IsConnectionKeptWhenClusterOverloaded test.

RB: 24253
Reviewed by: Andrzej Religa <andrzej.religa@oracle.com>