Bug#31316054 move connecting to backends to mysql_routing.cc
Problem
=======

RoutingDestination is manages the endpoints the routing plugin can
send traffic to.

To manage the state of the destinations its get_server_socket()
function connect()s to the backend, handles possible errors, switches to
the next destination ... until its successful.

Managing the state and actively connecting in the same class makes
changing the socket implementation hard.

Tracking the state of the destinations and connection to the
destinations should be split.

Change
======

- added a generic Destination class which maintains the state of
  which destinations are candidates, and in what order

Additionally,

- refactored available-destinations
- removed routingsockops, get_server_socket as they are unused now.
- removed unused include headers
- removed unused "using" declarations

RB: 24418