Bug#31031886: EVALUATING REGEXP CHARACTER CLASSES IN BINARY   STRING CONTEXTS
Bug#31031888: REGEXP_REPLACE AND REGEXP_SUBSTR STILL ADD NULL
  BYTES

Users get surprising effects when using binary strings as
arguments to regular expression functions, mainly because of
character classes. Internally the strings get mapped to
UTF-16 in a somewhat predictiable fashion, but this is such
an uncommon character set in the MySQL realm that the
results appear totally random.

We fix this by interpreting binary strings as cp-1252
instead (this is known as latin1 in MySQL). The point is
that each byte value is equal only to itself. But this is
really a special case; therefore we enforce that if one
argument is a "binary" string, then all of them have to be.

We also make sure to convert back to cp-1252, but calling
it binary, this fixes the problem with adding NULL bytes.

Change-Id: I3b27a26d660da8f77eeedd01df583f98880c5ea1