Bug #30620011 PB2 INNODB.INNODB_REDO_DEBUG_6 FAILS SPORADICALLY
Problem:

When the btr_mtr is committed and restored, the structure of the B-tree can
change.  Because of this the latches of internal nodes could be taken.
This case is not handled in the lob::purge() code.

Solution:

One solution is to avoid the intermediate commit/restore of the btr_mtr.  To do
that we have to ensure that conflicting changes are not done to btr_mtr and
lob_mtr.  The idea is to avoid conflicting changes in btr_mtr and the lob_mtr.

This is done by reserving extents after LOB operation.

rb#24314 in review by Debarun Banerjee <debarun.banerjee@oracle.com>