Bug #31045160 	TEST TRUNCATE_RECOVER_E09 FAILS SPORADICALLY ON RESTART
The first attempt to fix this test case  by adding more time to restart did not
work.  This testcase is still failing periodically on PB2.

It is easy to reproduce with this command:

./mtr portability_tablespace --parallel=10 --repeat=50

The failures show an error log that contains messages about page corruption
on disk. It occurs at startup after the intentional crash which for this test
is just after undo truncation completes.  The error log indicates that after
a few of the rollback segment header pages have been read without error, the
rest of them are corrupted.

The call to set_undo_tablespace_encryption() inside trx_undo_truncate_tablespace()
is located after the rollback segment header pages are created, meaning that
they are initially created without encryption and then encryption is turned on
while those pages are being whitten.

This fix moves the call to set_undo_tablespace_encryption() higher in
trx_undo_truncate_tablespace() so that it occurs within the same mtr as
the creation of the header page.  This way, every rollback segment header
page is written encrypted.

Approved by Mayank in RB#24344.