Bug #30716037 MEMORY CORRUPTION BUG IN MYSQLD
The temptable::Handler's implementation of records_in_range was not considered ready, so it contained DBUG_ABORT and return 0.
This lead to assertion failures in debug mode, and to returning "Empty set" in release mode for some queries.
The default implementation of records_in_range in the base class is to return 10, and while not ideal, is good enough.
Returning 0 means that we are sure that there are no records.
Returning 10 means that we estimate that there are 10 records.
The first statement might be wrong, the second is fuzzy enough, that it's technically never wrong.
This patch simply removes the overriden version of records_in_range from temptable::Handler.

Reviewed by: Jusufadis Bakamovic <jusufadis.bakamovic@oracle.com>
Reviewed by: Roy Lyseng <roy.lyseng@oracle.com>
RB:24321