Bug#28651673-SIMPLE FUNCTION CALLED IN SUBQUERY CRASHING MYSQL
Cause:

Tables used by a "Item_func" are represented by a bit map of the format
xxx_TABLE_BIT. The three pseudo bits are used for book-keeping and the
remaining bits represent the actual tables.

In this case of a sub-query with a deterministic function, the pseudo
bit is set to prevent the function from being evaluated in the
preparation phase. The function is to be evaluated later after the locks
are taken.

This pseudo table bit is wrongly assumed to be a bit representing an
actual table and when its corresponding table entry in the join table
array is accessed, the server exits if array entry does not exist.

Fix:

Skip the pseudo table bits while searching for tables through the
used-tables bit map.

Change-Id: I21a0bc9fabe18e3e13140594b505a41d851a88c8