Bug #31019990 : NDB 8.0 UPGRADE : MGMD CAN SEND INCORRECT LONG STOP_REQ TO OLDER NODES
Problem:
--------
1)  Start < 8.0.18 version cluster
2)  Upgrade MGMD(s) to >= 8.0.18
3)  Upgrade Master NDBD to >= 8.0.18
4)  Issue STOP with multiple node ids for < 8.0.18 data nodes
5)  The new master(< 8.0.18) crashes

The crash is because of wrong setting of signal length and number of signal
sections when sending GSN_STOP_REQ to the new master.
The wrong setting is due to the stale knowledge from sending GSN_STOP_REQ
to the old master.
Test coverage comes from:
 testUpgrade -n ChangeMGMDChangeHalfRestartChangeHalf

Solution:
--------
Set the signal length and number of sections every time before sending GSN_STOP_REQ
in MGMD

Reviewed-by: Frazer Clement <frazer.clement@oracle.com>