Bug #31252625: DATA IN WRONG ROW ON LEFT JOIN
A refactoring in bug #30460528 (RENAME FIELD::REAL_MAYBE_NULL() TO FIELD::IS_NULLABLE())
caused copying (single-row buffering for GROUP BY) of non-nullable fields to be
broken; it did not properly take into account that they could be the inner table
of an outer join, and thus would have a NULL flag that would need to be copied.
In GROUP BY without a temporary table, this would make the NULL flag come from
the _next_ output row instead of the previous one, and the data be somewhat random.

Change-Id: I065e75fc539e7aa6dbcdbc47545b207bb0dac291